+++ src/processor/modal.ts
@@ -4,6 +4,7 @@ import {getRandomQuoteId} from "../util/random";
 import LocalQuotes from "../main";
 import {parseBlockMetadataToCodeBlock, parseTime} from "../util/parser";
 import {createDomLink} from "../util/dom";
+import {getValidAuthorsFromAdvancedSearch} from "../types/quote";
 
 function fetchAuthorsInQuoteVault(plugin: LocalQuotes): Array<string> {
 	return plugin.quoteVault.map((obj) => obj.author);
@@ -53,8 +54,7 @@ export class QuoteMakerModal extends Modal {
 		const randomId: string = getRandomQuoteId();
 		let tmpRefreshChar: string = 'd';
 		let tmpRefreshNum: Number = 1;
-		let tmpSearch = '';
-		let tmpAuthor = authorList[0];
+		let tmpSearch = authorList[0];
 		this.result = {
 			content: {author: null, text: null},
 			id: randomId, search: null, refresh: 86400, customClass: null, lastUpdate: 0
@@ -85,15 +85,21 @@ export class QuoteMakerModal extends Modal {
 				}));
 
 		contentEl.createEl('h3', {text: 'Search'});
+		contentEl.createEl('p',
+			{
+				text: '(in the quote block will be used a last edited field (\'Author\' or \'Advanced search\'))',
+				cls: 'local-quote-heading-desc'
+			}
+		);
 
 		new Setting(contentEl)
 			.setName('Author')
 			.setDesc('Choose author for your quote block (if you want choose only one)')
 			.addDropdown(dropdown => {
 				authorList.map((author) => dropdown.addOption(author, author));
-				dropdown.setValue(tmpAuthor);
+				dropdown.setValue(tmpSearch);
 				dropdown.onChange((value) => {
-					tmpAuthor = value;
+					tmpSearch = value;
 				})
 			});
 
@@ -105,13 +111,23 @@ export class QuoteMakerModal extends Modal {
 					p.appendText('You can freely use \'*\' and || after reading ');
 					p.appendChild(createDomLink(p, 'search guide',
 						'https://github.com/ka1tzyu/local-quotes/wiki/How-to-use-search'));
-					p.appendText('.');
 				})
 			)
 			.addText(text => text
 				.setValue(tmpSearch)
 				.onChange((value) => {
 					tmpSearch = value;
+				}));
+
+		new Setting(contentEl)
+			.setName('Advanced search validator')
+			.setDesc('Summons notice with validated (existent) authors')
+			.addButton(btn => btn
+				.setButtonText('Check')
+				.setCta()
+				.onClick(() => {
+					new Notice('Validated authors:\n' +
+						getValidAuthorsFromAdvancedSearch(this.plugin, tmpSearch).join('\n'));
 				}));
 
 
@@ -161,7 +177,7 @@ export class QuoteMakerModal extends Modal {
 				.setButtonText('Insert Quote')
 				.setCta()
 				.onClick(() => {
-					this.result.search = tmpSearch.length > 0 ? tmpSearch : tmpAuthor;
+					this.result.search = tmpSearch;
 
 					this.editor.replaceRange(
 						parseBlockMetadataToCodeBlock(this.result, tmpRefreshNum + tmpRefreshChar),